# Ecommerce-swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Ecommerce - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: ndt11102003@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://ecommerce-market-api.vercel.app
...
paths:  
...
  /users/get-user/{id}:
    get:
      tags:
        - users
      summary: Get user by id
      security:
        - BearerAuth: []
      operationId: getUserByid
      parameters:
        - name: id
          in: path
          required: true
          description: ID of user to return
          schema:
            type: string
      responses:
        '200':
          description: Get user by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
  /users/update-password:
     post:
      tags:
        - users
      summary: Update password user 
      operationId: Update password
      requestBody:
        description: Update password field
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePassword'
              required:
                - oldPassword
                - newPassword
                - confirmNewPassword
      responses:
        '200':
          description: Update password successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Update password successfully"
                  status:
                    type: number
                    example: 200
                  
        '400':
          description: Update password failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
...
  /users/reset-password/{token}:
    put:
      tags:
        - users
      summary: Reset password
      operationId: resetpassword
      security:
        - BearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          description: Token to reset password
          schema:
            type: string
      requestBody:
        description: Check email to get token
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "password reset"
                confirmPassword:
                  type: string
                  example: "password reset"
      responses:
        '200':
          description: Reset password successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reset password successfully"
                  status:
                    type: number
                    example: 200
        '400':
          description: Forgot password failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: "64e8bd5a94a50c038406e72e"
        firstname:
          type: string
          example: "Nguyễn"
        lastname:
          type: string
          example: "Tân"
        email:
          type: string
          example: "ndt11102003@gmail.com"
        mobile:
          type: string
          example: "39123198"
        role:
          type: string
          example: "admin"
          enum:
            - admin
            - user
        address:
          type: string
          example: ""
        blocked:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2023-08-25T14:40:26.960Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-08-30T14:50:50.860Z"
        wishlist:
          type: array
          items:
            type: string
            format: ObjectId
            example: "650596a4f95fbf111e2f1e3e"
        avatar:
          type: string
          format: uri
          example: "https://cdn-icons-png.flaticon.com/512/2304/2304226.png"
    UpdatePassword:
      type: object
      properties:
        oldPassword:
          type: string
          example: "password old"
        newPassword:
          type: string
          example: "Admin123"
        confirmNewPassword:
          type: string
          example: "Admin123"
    ResponseError:
      type: object
      properties:
        error:
          type: string
          example: "Invalid value"
        path:
          type: string
          example: "path field error"
        status:
          type: integer
          example: 400
...
  securitySchemes:
      BearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT